RS_Entity: ID type to unsigned long long
This helps avoid running out of ID numbers

+++ librecad/src/lib/engine/rs_entity.cpp
@@ -97,7 +97,7 @@ void RS_Entity::init() {
  * Gives this entity a new unique id.
  */
 void RS_Entity::initId() {
-    static unsigned long int idCounter=0;
+    static unsigned long long idCounter=0;
     id = idCounter++;
 }
